moved raspUtils into dasBidAdapter.js
